+++ libraries/provider_vlc_app_mac_os_x.rb
@@ -75,8 +75,16 @@ class Chef
         # @return [String] a download URL
         #
         def remote_path
-          "https://get.videolan.org/vlc/#{latest_version}/macosx/" \
-            "vlc-#{latest_version}.dmg"
+          "https://get.videolan.org/vlc/#{version}/macosx/vlc-#{version}.dmg"
+        end
+
+        #
+        # Return either the new_resource's version or get the latest one.
+        #
+        # @return [String] a version string for this provider to use
+        #
+        def version
+          new_resource.version || latest_version
         end
       end
     end