+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Vlc Cookbook CHANGELOG
 
 v?.?.? (????-??-??)
 -------------------
+- Give the `vlc_app` resource an optional `version` attribute
 
 v0.2.0 (2015-06-21)
 -------------------